pyproject.toml
[tool.poetry]
name = "tools-api"
version = "0.1.0"
description = ""
authors = ["Sumit Guha <sumit03guha@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
flask = "3.0.3"
flask-restx = "1.3.0"
pymongo = "4.8.0"
python-dotenv = "1.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "8.2.2"
pytest-flask = "1.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
